Moving linkRewriter logic inside amp-skimlinks
