+++ extensions/amp-skimlinks/0.1/link-rewriter/event-messenger.js
