+++ extensions/amp-skimlinks/0.1/link-rewriter/link-replacement-cache.js
