+++ src/service/navigation.js
@@ -47,6 +47,7 @@ const ORIG_HREF_ATTRIBUTE = 'data-a4a-orig-href';
 /** @enum {number} */
 export const Priority = {
   ANALYTICS_LINKER: 2,
+  LINK_REWRITER_MANAGER: 3,
 };
 
 /**
@@ -100,9 +101,6 @@ export class Navigation {
 
     /** @private @const {!./history-impl.History} */
     this.history_ = Services.historyForDoc(this.ampdoc);
-
-    /** @private @const {!./link-rewriter/link-rewriter-manager.LinkRewriterManager} */
-    this.linkRewriterService_ = Services.linkRewriterServiceForDoc(this.ampdoc);
 
     const platform = Services.platformFor(this.ampdoc.win);
     /** @private @const {boolean} */
@@ -257,10 +255,6 @@ export class Navigation {
     if (!target || !target.href) {
       return;
     }
-    this.linkRewriterService_.maybeRewriteLink(
-        /** @type {!HTMLElement} */ (target),
-        e.type
-    );
 
     if (e.type == EVENT_TYPE_CLICK) {
       this.handleClick_(target, e);
